Extract App from models package
Final model to extract for consistent structure of all model packages.
